Only exercise tiers the host can execute; fix trampoline stackmap
The conformance suite ran every tier compiled into the binary, which is the
point — it lets one process compare avx2 against avx512 against the reference
without re-running per GOSIMD setting. But a tier the host cannot execute is
not a failing test, it is a SIGILL with no message. It died on the first
vsetvli under a riscv64 emulator with no vector extension, and would do the
same to anyone running the suite on an amd64 machine older than AVX-512.

The library's own dispatch was correct throughout and would have selected the
scalar tier; only the test reached past what the CPU advertised. Skipped tiers
are now logged, because a suite that silently tested nothing looks exactly
like one that passed.

Also: a framed assembly function needs to declare whether its locals hold
pointers, or the runtime aborts with 'missing stackmap' the first time a
goroutine stack grows under one. The save-area trampoline now emits
NO_LOCAL_POINTERS.

Substring search (Bytes.Index) is vectorized: compare a register against
broadcasts of the needle's first and last bytes and verify only the positions
where both match, which on random text rejects all but one in 65536 without
touching the middle of the needle. Bytes is now 21 of 24.

+++ simd_test.go
@@ -86,10 +86,16 @@ func TestElementwiseMatchesScalarLoop(t *testing.T) {
 		check("Mul", dst, func(x, y float64) float64 { return x * y })
 		simd.DivInto(dst, a, b)
 		check("Div", dst, func(x, y float64) float64 { return x / y })
+		// The conversions around the products are load-bearing on any
+		// architecture where Go fuses a multiply into a following add — arm64,
+		// ppc64, s390x and riscv64 all do, amd64 does not. Without them this
+		// expectation is computed with one rounding where the kernel uses two,
+		// and the test fails by an ULP on four of the six architectures while
+		// passing on the machine it was written on.
 		simd.AddScaledInto(dst, a, b, 2.5)
-		check("AddScaled", dst, func(x, y float64) float64 { return x + y*2.5 })
+		check("AddScaled", dst, func(x, y float64) float64 { return x + float64(y*2.5) })
 		simd.LerpInto(dst, a, b, 0.25)
-		check("Lerp", dst, func(x, y float64) float64 { return x + (y-x)*0.25 })
+		check("Lerp", dst, func(x, y float64) float64 { return x + float64((y-x)*0.25) })
 	}
 }
